Axon now detects imminent deadlocks in the IdentifierBasedLock
Before attempting to acquire a lock, the thread checks whether the lock is
already owned by a thread waiting for a lock owned by the current thread. If
that is the case, the lock is not acquired and a DeadlockException is thrown.
This will release the other locks, giving the other thread the opportunity to
acquire the lock it was waiting for and continue processing. The failed
operation can usually be retried automatically.

Issue #AXON-13 add Fix versions 1.5